requirements.txt
# SQL Query Optimizer - Dependencies

# Core Dependencies
psycopg2-binary>=2.9.0
sqlparse>=0.4.0
pandas>=1.3.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# CLI and Formatting  
tabulate>=0.9.0

# System utilities
python-dateutil>=2.8.0

# Development and Testing (optional)
pytest>=6.0.0
